add tests for syntax highlighter
